+++ src/__tests__/Microcosm-test.jsx
@@ -1,6 +1,8 @@
 let Action     = require('./fixtures/Action')
 let DummyStore = require('./fixtures/DummyStore')
 let Microcosm  = require('../Microcosm')
+let React      = require('react/addons')
+let render     = React.addons.TestUtils.renderIntoDocument
 
 describe('Microcosm', function() {
   let app;
@@ -200,8 +202,8 @@ describe('Microcosm', function() {
   describe('::start', function() {
     it ('can run multiple callbacks', function(done) {
       let app = new Microcosm()
-      let a   = sinon.mock()
-      let b   = sinon.mock()
+      let a   = sinon.stub()
+      let b   = sinon.stub()
 
       app.start(a, b, function() {
         a.should.have.been.called
@@ -219,6 +221,13 @@ describe('Microcosm', function() {
         error.should.be.instanceof(TypeError)
         error.message.should.include('start')
       }
+    })
+  })
+
+  describe('::render', function() {
+    it ('renders a Microscope', function() {
+      let el = render(<Microcosm instance={ app }><p></p></Microcosm>)
+      el.props.instance.should.equal(app)
     })
   })